Add files via upload
R script and initial R data frame for the analysis of a selection of some 120 variables from cumulative data of the European Social Survey (ESS), rounds 1 to 9. Details and references to the ESS sources are given in the initial part of the R script. The script includes the R code developed for a preliminary coding of political parties as far right-wing political parties, in which work from political science is used as coding sources (references cited in the scipt).